Implement CRUD operations for product management, add 4 new models, and integrate search, sorting, pagination and get list functionality

+++ Backend/computer_accessories/src/main/java/com/dev/computer_accessories/repository/SupplierRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface SupplierRepository extends JpaRepository<Supplier, Integer> {
+public interface SupplierRepository extends JpaRepository<Supplier, Long> {
 
 
     Optional<Supplier> findByName(String name);
